Refactor contact form to separate components for add and edit screens
