+++ packages/client/screens/contacts/EditContactScreen.tsx
@@ -18,292 +18,21 @@ import { getCurrencyType, isValidCurrencyAddress } from "@cryptify/common/src/ut
 import { ERROR_WALLET_ADDRESS_INVALID_FOR_CURRENCY } from "@cryptify/common/src/errors/error_messages";
 import { ContactsGateway } from "../../gateways/contacts_gateway";
 import { AuthContext } from "../../components/contexts/AuthContext";
-import { CreateContactRequest } from "@cryptify/common/src/requests/create_contact_request";
 import { CompositeScreenProps } from "@react-navigation/native";
-
-type CreateContactRequestPayload = {
-    contactName: string,
-    userId: number,
-    ethWallets: string[],
-    btcWallets: string[],
-    ethWalletsDelete: string[],
-    btcWalletsDelete: string[],
-}
+import {CreateContactRequest} from "@cryptify/common/src/requests/create_contact_request";
+import ContactsForm from "../../components/contacts/ContactsForm";
 
 export default function EditContactScreen(props: SettingsStackScreenProps<"EditContactScreen">) {
-    const contactsGateway = new ContactsGateway();
-    const isEditMode = !!props.route.params?.contact;
-
-    const { token, user } = React.useContext(AuthContext);
-
-    const defaultEthWallets = isValidCurrencyAddress(
-        props.route.params?.prefilledWalletAddress || "",
-        CurrencyType.ETHEREUM,
-    )
-        ? [props.route.params?.prefilledWalletAddress || ""]
-        : isEditMode
-        ? props.route.params.contact.addresses
-              .filter((c) => getCurrencyType(c.walletAddress) === CurrencyType.ETHEREUM)
-              .map((c) => c.walletAddress)
-        : [];
-    const defaultBtcWallets = isValidCurrencyAddress(
-        props.route.params?.prefilledWalletAddress || "",
-        CurrencyType.BITCOIN,
-    )
-        ? [props.route.params?.prefilledWalletAddress || ""]
-        : isEditMode
-        ? props.route.params.contact.addresses
-              .filter((c) => getCurrencyType(c.walletAddress) === CurrencyType.BITCOIN)
-              .map((c) => c.walletAddress)
-        : [];
-
-    const initialValues: CreateContactRequestPayload = {
-        contactName: isEditMode ? props.route.params.contact.contactName : "",
-        userId: user.id,
-        ethWallets: defaultEthWallets,
-        btcWallets: defaultBtcWallets,
-        ethWalletsDelete: [],
-        btcWalletsDelete: [],
-    };
-
-    function AddWalletFieldArray({
-        values,
-        handleChange,
-        currencyType,
-        errors,
-        touched,
-        placeholder,
-    }: AddWalletFieldArrayProps) {
-        const currencyIcon = currencyType === CurrencyType.BITCOIN ? faBitcoin : faEthereum;
-        const iconColor = currencyType === CurrencyType.BITCOIN ? "#F7931A" : "#3C3C3D";
-
-        const wallets = currencyType === CurrencyType.BITCOIN ? values.btcWallets : values.ethWallets;
-        const currencyErrors = currencyType === CurrencyType.BITCOIN ? errors.btcWallets : errors.ethWallets;
-        const currencyTouched = currencyType === CurrencyType.BITCOIN ? touched.btcWallets : touched.ethWallets;
-        const walletListString = currencyType === CurrencyType.BITCOIN ? "btcWallets" : "ethWallets";
-
-        const [isCollapsed, setIsCollapsed] = React.useState<boolean>(
-            !props.route.params?.prefilledWalletAddress && !isEditMode,
-        );
-
-        return (
-            <>
-                <Pressable
-                    onPress={() => {
-                        setIsCollapsed(!isCollapsed);
-                    }}
-                    testID="walletCollapsibleButton"
-                >
-                    <HStack style={{ marginTop: 40 }}>
-                        <FontAwesomeIcon style={{ marginRight: 10 }} color={iconColor} icon={currencyIcon} size={26} />
-                        <Text fontWeight={"semibold"} size={"title3"}>
-                            {titleCase(currencyType)} Wallets
-                        </Text>
-                        {isCollapsed ? (
-                            <FontAwesomeIcon style={styles.chevronIcon} size={18} icon={farChevronDown} />
-                        ) : (
-                            <FontAwesomeIcon style={styles.chevronIcon} icon={farChevronUp} size={18} />
-                        )}
-                    </HStack>
-                </Pressable>
-                {/* @ts-expect-error this is a known type error in the dependency a pr was made to address the issue but never merged */}
-                <Collapsible collapsed={isCollapsed}>
-                    <FieldArray
-                        name={walletListString}
-                        render={(arrayHelpers) => (
-                            <View>
-                                {wallets?.map((wallet, i) => (
-                                    <View style={wallets.length > 1 ? { marginTop: 13 } : { marginTop: 20 }} key={i}>
-                                        <FormControl
-                                            isInvalid={!!(currencyErrors?.[i])}
-                                        >
-                                            <Input
-                                                value={wallet}
-                                                onChangeText={handleChange(`${walletListString}[${i}]`)}
-                                                placeholder={placeholder}
-                                                rightElement={
-                                                    !props.route.params?.prefilledWalletAddress ? (
-                                                        <Pressable
-                                                            onPress={() => {
-                                                                arrayHelpers.remove(i);
-                                                            }}
-                                                        >
-                                                            <FontAwesomeIcon
-                                                                color={"#EF4444"}
-                                                                style={{ marginRight: 12 }}
-                                                                size={20}
-                                                                icon={falCircleXMark}
-                                                            />
-                                                        </Pressable>
-                                                    ) : (
-                                                        <></>
-                                                    )
-                                                }
-                                                isDisabled={!!props.route.params?.prefilledWalletAddress}
-                                                testID="walletAddressInput"
-                                            />
-                                            <FormControl.ErrorMessage>
-                                                {currencyErrors ? currencyErrors[i] : ""}
-                                            </FormControl.ErrorMessage>
-                                        </FormControl>
-                                    </View>
-                                ))}
-                                {!props.route.params?.prefilledWalletAddress && (
-                                    <View>
-                                        <Pressable
-                                            onPress={() => {
-                                                arrayHelpers.push("");
-                                            }}
-                                            testID={`addAnother${currencyType}`}
-                                        >
-                                            <HStack style={{ marginTop: 13 }}>
-                                                <FontAwesomeIcon
-                                                    color={"#0077E6"}
-                                                    icon={fasCirclePlusSolid}
-                                                    size={20}
-                                                />
-                                                <Text
-                                                    style={{ marginLeft: 10 }}
-                                                    color={"darkBlue.500"}
-                                                    fontWeight={"semibold"}
-                                                >
-                                                    Add another {titleCase(currencyType)} wallet
-                                                </Text>
-                                            </HStack>
-                                        </Pressable>
-                                    </View>
-                                )}
-                            </View>
-                        )}
-                    />
-                </Collapsible>
-            </>
-        );
-    }
-
-    async function onAddContactSubmit(
-        values: CreateContactRequestPayload,
-        formikHelpers: FormikHelpers<CreateContactRequestPayload>,
-    ) {
-        const currencies = [CurrencyType.BITCOIN, CurrencyType.ETHEREUM];
-        let hasError = false;
-
-        // checking for errors for each currency in the list
-        currencies.map((currencyType) => {
-            const wallets = currencyType === CurrencyType.BITCOIN ? values.btcWallets : values.ethWallets;
-            const walletListString = currencyType === CurrencyType.BITCOIN ? "btcWallets" : "ethWallets";
-
-            // validating each string to be a valid currency wallet
-            wallets?.map((walletAddress, i) => {
-                // ignoring empty strings
-                if (walletAddress !== undefined && walletAddress.length > 0) {
-                    const isAddressValid = isValidCurrencyAddress(walletAddress, currencyType);
-                    if (!isAddressValid) {
-                        formikHelpers.setFieldError(
-                            `${walletListString}[${i}]`,
-                            ERROR_WALLET_ADDRESS_INVALID_FOR_CURRENCY(titleCase(currencyType)).split(":")[1],
-                        );
-                        hasError = true;
-                        return;
-                    }
-                }
-            });
-        });
-
-        if (hasError) {
-            return;
-        }
-
-        try {
-            const walletAddrs = [...values.ethWallets, ...values.btcWallets].filter((addr) => addr.length !== 0);
-            await contactsGateway.createContact({
-                contactName: values.contactName,
-                userId: user.id,
-                walletAddrs,
-            }, token);
-            
-            props.navigation.goBack();
-            if (props.route.params?.prefilledWalletAddress) {
-                props.navigation.goBack();
-            }
-        } catch (error) {
-            if (error instanceof Error) {
-                formikHelpers.setFieldError("contactName", error.message);
-            }
-        }
-    }
-
     return (
         <View style={styles.view}>
-            <Formik initialValues={initialValues} onSubmit={onAddContactSubmit}>
-                {({ values, errors, touched, handleChange, submitForm }) => (
-                    <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
-                        <FormControl isInvalid={!!(errors.contactName && touched.contactName)}>
-                            <Input
-                                value={values.contactName}
-                                autoFocus={true}
-                                onChangeText={handleChange("contactName")}
-                                placeholder="Name"
-                                maxLength={20}
-                                keyboardType={"ascii-capable"}
-                                testID="contactNameInput"
-                            />
-                            <FormControl.ErrorMessage>{errors.contactName}</FormControl.ErrorMessage>
-                        </FormControl>
-                        {!props.route.params?.prefilledWalletAddress ? (
-                            <>
-                                <AddWalletFieldArray
-                                    values={values}
-                                    handleChange={handleChange}
-                                    currencyType={CurrencyType.BITCOIN}
-                                    errors={errors}
-                                    touched={touched}
-                                    placeholder={"Wallet address (Begins with 1, 3, or bc1)"}
-                                />
-                                <AddWalletFieldArray
-                                    values={values}
-                                    handleChange={handleChange}
-                                    currencyType={CurrencyType.ETHEREUM}
-                                    errors={errors}
-                                    touched={touched}
-                                    placeholder={"Wallet address (Begins with 0x)"}
-                                />
-                            </>
-                        ) : (
-                            <AddWalletFieldArray
-                                values={values}
-                                handleChange={handleChange}
-                                currencyType={getCurrencyType(props.route.params?.prefilledWalletAddress)}
-                                errors={errors}
-                                touched={touched}
-                            />
-                        )}
-                        <Button
-                            style={
-                                values.contactName.length > 0
-                                    ? styles.addContactButton
-                                    : styles.addContactButtonDisabled
-                            }
-                            onPress={submitForm}
-                            testID="submitCreateContactButton"
-                        >
-                            {isEditMode ? "Edit Contact" : "Add Contact"}
-                        </Button>
-                    </ScrollView>
-                )}
-            </Formik>
+            <ContactsForm
+                prefilledWalletAddress={undefined}
+                contact={props.route.params.contact}
+                navigation={props.navigation}
+            />
         </View>
     );
 }
-
-type AddWalletFieldArrayProps = {
-    values: CreateContactRequestPayload;
-    handleChange: any;
-    currencyType: CurrencyType;
-    errors: FormikErrors<CreateContactRequestPayload>;
-    touched: FormikTouched<CreateContactRequestPayload>;
-    placeholder?: string;
-};
 
 const styles = StyleSheet.create({
     view: {
@@ -311,16 +40,5 @@ const styles = StyleSheet.create({
         paddingHorizontal: 15,
         paddingTop: 30,
         paddingBottom: 15,
-    },
-    chevronIcon: {
-        marginLeft: "auto",
-        marginRight: 5,
-    },
-    addContactButton: {
-        marginTop: "auto",
-    },
-    addContactButtonDisabled: {
-        marginTop: "auto",
-        opacity: 0.6,
     },
 });
