+++ packages/common/src/utils/function_utils.ts
@@ -1,3 +1,13 @@
 export function zip<T, U>(a: T[], b: U[]): [T, U][] {
     return a.map((k, i) => [k, b[i]]);
 }
+
+export function equals<T>(a: T[], b: T[]): boolean {
+    if (a.length !== b.length) {
+        return false;
+    }
+    
+    const x = [...a].sort();
+    const y = [...b].sort();
+    return x.every((v, i) => v === y[i]);
+}
